Start splitting up Firestore functions into separate files
